requirements.txt
apache-airflow-providers-fab
apache-airflow-providers-celery
apache-airflow-providers-docker
apache-airflow-providers-postgres
google-cloud-bigquery
dbt-bigquery
python-dotenv